requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests>=2.31.0
feedparser>=6.0.11

# AI API clients
google-generativeai>=0.8.0
openai>=1.12.0
anthropic>=0.18.1

# PDF handling
PyPDF2==3.0.1
arxiv>=2.1.0

# Data storage
tinydb==4.8.0

# Web generation
jinja2==3.1.3
markdown==3.5.2

# Utilities
python-dateutil==2.8.2
beautifulsoup4==4.12.3
